Move d20 HTML generation to d20Roll
