+++ src/system/dice/d20-roll.ts
@@ -7,6 +7,7 @@ import { RollConfigurationDialog } from '@system/applications/dialogs/roll-confi
 
 import { PlotDie } from './plot-die';
 import { RollMode } from './types';
+import { hasKey, renderSystemTemplate, TEMPLATES } from '../utils';
 
 // Constants
 const CONFIGURATION_DIALOG_TEMPLATE =
@@ -290,6 +291,112 @@ export class D20Roll extends foundry.dice.Roll<D20RollData> {
         options.rollMode ??= game.settings!.get('core', 'rollMode');
 
         return super.toMessage(messageData, options);
+    }
+
+    public async getHTML() {
+        const OPPORTUNITY = 'opportunity';
+        const COMPLICATION = 'complication';
+
+        if (!this.validD20Roll) return;
+
+        // Process bonuses beyond the base d20s into a single roll.
+        const bonusTerms = this.terms.slice(1);
+
+        for (const term of bonusTerms) {
+            // Terms throw an error if already evaluated. We can ignore them if so.
+            try {
+                await term.evaluate();
+            } catch (err) {
+                continue;
+            }
+        }
+
+        const bonusRoll =
+            bonusTerms && bonusTerms.length > 0
+                ? Roll.fromTerms(bonusTerms)
+                : null;
+        const d20Dice = this.dice.find((d) => d.faces === 20);
+
+        if (!d20Dice) return;
+
+        const plot = [];
+
+        if (this.hasPlotDie) {
+            const plotDice = this.terms.filter((r) => r instanceof PlotDie);
+            for (const plotDie of plotDice) {
+                if (plotDie.rolledOpportunity) plot.push(OPPORTUNITY);
+                if (plotDie.rolledComplication) plot.push(COMPLICATION);
+            }
+        }
+
+        const entries = [];
+        for (let i = 0; i < d20Dice.results.length; i++) {
+            const tmpResults = [];
+            tmpResults.push(foundry.utils.duplicate(d20Dice.results[i]));
+
+            while (
+                d20Dice?.results[i]?.rerolled &&
+                !d20Dice?.results[i]?.count
+            ) {
+                if (i + 1 >= d20Dice.results.length) {
+                    break;
+                }
+
+                i++;
+                tmpResults.push(foundry.utils.duplicate(d20Dice.results[i]));
+            }
+
+            // Die terms must have active results or the base roll total of the generated roll is 0.
+            // This does not apply to dice that have been rerolled (unless they are replaced by a fixer value eg. for reliable talent).
+            tmpResults.forEach((r) => {
+                r.active = !(r.rerolled && !r.count);
+            });
+
+            const modifiers = new Array<
+                keyof (typeof foundry.dice.terms.Die)['MODIFIERS']
+            >();
+            for (const mod of d20Dice.modifiers) {
+                if (hasKey(foundry.dice.terms.Die.MODIFIERS, mod)) {
+                    modifiers.push(mod);
+                }
+            }
+
+            const baseTerm = new foundry.dice.terms.Die({
+                number: 1,
+                faces: 20,
+                results: tmpResults,
+                modifiers,
+            });
+            const baseRoll = D20Roll.fromTerms([baseTerm]);
+
+            const total = (baseRoll?.total ?? 0) + (bonusRoll?.total ?? 0);
+
+            const plotD20 = [...plot];
+            for (let o = 0; o < baseRoll.opportunitiesCount; o++) {
+                plotD20.push(OPPORTUNITY);
+            }
+            for (let c = 0; c < baseRoll.complicationsCount; c++) {
+                plotD20.push(COMPLICATION);
+            }
+
+            entries.push({
+                roll: baseRoll,
+                total: total,
+                ignored: tmpResults.some((r) => r.discarded) ? true : undefined,
+                plotType: plotD20.some((p) => p === OPPORTUNITY)
+                    ? OPPORTUNITY
+                    : plotD20.some((p) => p === COMPLICATION)
+                      ? COMPLICATION
+                      : undefined,
+                plotDice: plotD20,
+            });
+        }
+
+        return renderSystemTemplate(TEMPLATES.CHAT_ROLL_D20, {
+            formula: this.formula,
+            tooltip: await this.getTooltip(),
+            entries,
+        });
     }
 
     /* --- Internal Functions --- */
