+++ src/system/documents/chat-message.ts
@@ -3,8 +3,7 @@ import { D20Roll } from '@system/dice/d20-roll';
 import { DamageRoll } from '@system/dice/damage-roll';
 
 import { CosmereActor } from './actor';
-import { hasKey, renderSystemTemplate, TEMPLATES } from '../utils';
-import { PlotDie } from '../dice';
+import { renderSystemTemplate, TEMPLATES } from '../utils';
 
 const ACTIVITY_CARD_TEMPLATE =
     'systems/cosmere-rpg/templates/chat/activity-card.hbs';
@@ -105,7 +104,6 @@ export class CosmereChatMessage extends ChatMessage {
         const content = $(
             await renderSystemTemplate(TEMPLATES.CHAT_CARD_CONTENT, {}),
         );
-        content.find('.chat-card').append(html.find('.dice-roll'));
 
         await this.enrichSkillTest(content);
         await this.enrichDamage(content);
@@ -138,15 +136,14 @@ export class CosmereChatMessage extends ChatMessage {
 
         html.find('.chat-card').append(sectionHTML);
 
-        const roll = html.find('.dice-roll');
+        const rollHTML = await d20Roll.getHTML();
+        const roll = $(rollHTML as unknown as HTMLElement);
         const tooltip = roll.find('.dice-tooltip');
-        this.enrichD20Tooltip(this.d20Rolls[0], tooltip[0]);
+        this.enrichD20Tooltip(d20Roll, tooltip[0]);
         tooltip.prepend(roll.find('.dice-formula'));
 
         const section = html.find('.chat-card-section.skill');
         section.append(roll);
-
-        await this.enrichD20(d20Roll, html);
     }
 
     protected async enrichDamage(html: JQuery) {
@@ -162,113 +159,6 @@ export class CosmereChatMessage extends ChatMessage {
         );
 
         html.find('.chat-card').append(sectionHTML);
-    }
-
-    protected async enrichD20(roll: D20Roll, html: JQuery) {
-        const OPPORTUNITY = 'opportunity';
-        const COMPLICATION = 'complication';
-
-        if (!roll) return;
-
-        // Process bonuses beyond the base d20s into a single roll.
-        const bonusTerms = roll.terms.slice(1);
-
-        for (const term of bonusTerms) {
-            // Terms throw an error if already evaluated. We can ignore them if so.
-            try {
-                await term.evaluate();
-            } catch (err) {
-                continue;
-            }
-        }
-
-        const bonusRoll =
-            bonusTerms && bonusTerms.length > 0
-                ? Roll.fromTerms(bonusTerms)
-                : null;
-        const d20Dice = roll.dice.find((d) => d.faces === 20);
-
-        if (!d20Dice) return;
-
-        const plot = [];
-
-        if (roll.hasPlotDie) {
-            const plotDice = roll.terms.filter((r) => r instanceof PlotDie);
-            for (const plotDie of plotDice) {
-                if (plotDie.rolledOpportunity) plot.push(OPPORTUNITY);
-                if (plotDie.rolledComplication) plot.push(COMPLICATION);
-            }
-        }
-
-        const entries = [];
-        for (let i = 0; i < d20Dice.results.length; i++) {
-            const tmpResults = [];
-            tmpResults.push(foundry.utils.duplicate(d20Dice.results[i]));
-
-            while (
-                d20Dice?.results[i]?.rerolled &&
-                !d20Dice?.results[i]?.count
-            ) {
-                if (i + 1 >= d20Dice.results.length) {
-                    break;
-                }
-
-                i++;
-                tmpResults.push(foundry.utils.duplicate(d20Dice.results[i]));
-            }
-
-            // Die terms must have active results or the base roll total of the generated roll is 0.
-            // This does not apply to dice that have been rerolled (unless they are replaced by a fixer value eg. for reliable talent).
-            tmpResults.forEach((r) => {
-                r.active = !(r.rerolled && !r.count);
-            });
-
-            const modifiers = new Array<
-                keyof (typeof foundry.dice.terms.Die)['MODIFIERS']
-            >();
-            for (const mod of d20Dice.modifiers) {
-                if (hasKey(foundry.dice.terms.Die.MODIFIERS, mod)) {
-                    modifiers.push(mod);
-                }
-            }
-
-            const baseTerm = new foundry.dice.terms.Die({
-                number: 1,
-                faces: 20,
-                results: tmpResults,
-                modifiers,
-            });
-            const baseRoll = D20Roll.fromTerms([baseTerm]);
-
-            const total = (baseRoll?.total ?? 0) + (bonusRoll?.total ?? 0);
-
-            const plotD20 = [...plot];
-            for (let o = 0; o < baseRoll.opportunitiesCount; o++) {
-                plotD20.push(OPPORTUNITY);
-            }
-            for (let c = 0; c < baseRoll.complicationsCount; c++) {
-                plotD20.push(COMPLICATION);
-            }
-
-            entries.push({
-                roll: baseRoll,
-                total: total,
-                ignored: tmpResults.some((r) => r.discarded) ? true : undefined,
-                plotType: plotD20.some((p) => p === OPPORTUNITY)
-                    ? OPPORTUNITY
-                    : plotD20.some((p) => p === COMPLICATION)
-                      ? COMPLICATION
-                      : undefined,
-                plotDice: plotD20,
-            });
-        }
-
-        const rollHTML = await renderSystemTemplate(
-            TEMPLATES.CHAT_CARD_ROLL_D20,
-            { entries },
-        );
-
-        html.find('.dice-total').replaceWith(rollHTML);
     }
 
     /**
